requirements.txt
pydantic==1.8.2
python-dotenv==0.19.2
openpyxl>=3.1.0
pandas>=2.0.0
numpy>=1.24.0
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
PyPDF2>=3.0.0
requests>=2.31.0
boto3==1.26.137
openai==1.5.0
langchain==0.0.325
langchain-openai==0.0.2.post1
langchain-community==0.0.13
pypdf==3.17.1
pinecone-client==2.2.4
weaviate-client==3.25.2
crewai==0.11.2
jinja2==3.1.2
weasyprint==60.1
email-validator==2.1.0.post1
python-docx==0.8.11
pdfplumber==0.10.3
python-magic==0.4.27
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
Pillow==10.2.0
opencv-python==4.9.0.80
ezdxf==1.1.3
python-magic-bin==0.4.14; sys_platform == 'win32'
redis==5.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
slowapi==0.1.4
sendgrid==6.9.0
google-auth-oauthlib==0.4.6
google-auth-httplib2==0.1.0
google-api-python-client==2.31.0
pypdf==3.15.1
python-docx==0.8.11
pdfplumber==0.7.0
python-magic==0.4.27
opencv-python==4.5.3.56
numpy==1.21.2
pillow==8.3.2
ezdxf==0.17.2
pytest==6.2.5
pytest-cov==2.12.1
black==21.7b0
flake8==3.9.2
mypy==0.910 